Refactor the MTP and Eagle code
